# added faster dynamodb item deserializer