--- a/benchmarks/deserialize/README.md
+++ b/benchmarks/deserialize/README.md
@@ -1,0 +1,5 @@
+Run:
+
+* `poetry run  python bench_boto3.py -o boto3.json --rigorous`
+* `poetry run python bench_aiodynamo.py -o aiodynamo.json --rigorous`
+* `poetry run python -m pyperf compare boto3.json aiodynamo.json`