--- a/pyproject.toml
+++ b/pyproject.toml
@@ -34,12 +34,6 @@ multi_line_output = "3"
 combine_as_imports = "1"
 include_trailing_comma = "True"
 
-[tool.isort]
-line_length = "88"
-multi_line_output = "3"
-combine_as_imports = "1"
-include_trailing_comma = "True"
-
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"