--- a/tests/unit/test_utils.py
+++ b/tests/unit/test_utils.py
@@ -1,9 +1,10 @@
 from decimal import Decimal
+from functools import partial
 
 import pytest
 from aiodynamo import utils
 from aiodynamo.utils import clean, deserialize, dy2py, remove_empty_strings
-from boto3.dynamodb.types import DYNAMODB_CONTEXT
+from boto3.dynamodb.types import DYNAMODB_CONTEXT, TypeDeserializer
 
 
 @pytest.mark.asyncio
@@ -66,3 +67,37 @@ def test_numeric_decode(value, numeric_type, result):
 )
 def test_remove_empty_strings(item, result):
     assert dict(remove_empty_strings(item)) == result
+
+
+def test_fast_decode_compatibility():
+    def generate_item(nest):
+        item = {
+            "hash": {"S": "string",},
+            "range": {"B": b"bytes",},
+            "null": {"NULL": True},
+            "true": {"BOOL": True},
+            "false": {"BOOL": False},
+            "int": {"N": "42"},
+            "float": {"N": "4.2"},
+            "numeric_set": {"NS": ["42", "4.2"]},
+            "string_set": {"SS": ["hello", "world"]},
+            "binary_set": {"BS": [b"hello", b"world"]},
+        }
+        if nest:
+            item["list"] = {"L": [{"M": generate_item(False)}]}
+        return item
+
+    item = generate_item(True)
+
+    class BinaryDeserializer(TypeDeserializer):
+        def _deserialize_b(self, value):
+            return value
+
+    def deserialize_item(item, deserializer):
+        return {k: deserializer(v) for k, v in item.items()}
+
+    fast = deserialize_item(
+        item, partial(deserialize, numeric_type=DYNAMODB_CONTEXT.create_decimal)
+    )
+    boto = deserialize_item(item, BinaryDeserializer().deserialize)
+    assert fast == boto