local results from the benchmarks:

```
❯ poetry run python -mpyperf stats aiodynamo.json
Total duration: 2 min 14.1 sec
Start date: 2020-02-18 15:29:37
End date: 2020-02-18 15:32:47
Raw value minimum: 709 ms
Raw value maximum: 942 ms

Number of calibration run: 1
Number of run with values: 40
Total number of run: 41

Number of warmup per run: 1
Number of value per run: 3
Loop iterations per value: 1
Total number of values: 120

Minimum:         709 ms
Median +- MAD:   871 ms +- 35 ms
Mean +- std dev: 839 ms +- 76 ms
Maximum:         942 ms

  0th percentile: 709 ms (-16% of the mean) -- minimum
  5th percentile: 717 ms (-15% of the mean)
 25th percentile: 747 ms (-11% of the mean) -- Q1
 50th percentile: 871 ms (+4% of the mean) -- median
 75th percentile: 899 ms (+7% of the mean) -- Q3
 95th percentile: 934 ms (+11% of the mean)
100th percentile: 942 ms (+12% of the mean) -- maximum

Number of outlier (out of 521 ms..1125 ms): 0
❯ poetry run python -mpyperf stats boto3.json
Total duration: 4 min 29.3 sec
Start date: 2020-02-18 14:59:40
End date: 2020-02-18 15:05:06
Raw value minimum: 1.51 sec
Raw value maximum: 2.03 sec

Number of calibration run: 1
Number of run with values: 40
Total number of run: 41

Number of warmup per run: 1
Number of value per run: 3
Loop iterations per value: 1
Total number of values: 120

Minimum:         1.51 sec
Median +- MAD:   1.61 sec +- 0.06 sec
Mean +- std dev: 1.65 sec +- 0.11 sec
Maximum:         2.03 sec

  0th percentile: 1.51 sec (-9% of the mean) -- minimum
  5th percentile: 1.53 sec (-7% of the mean)
 25th percentile: 1.57 sec (-5% of the mean) -- Q1
 50th percentile: 1.61 sec (-3% of the mean) -- median
 75th percentile: 1.75 sec (+6% of the mean) -- Q3
 95th percentile: 1.82 sec (+10% of the mean)
100th percentile: 2.03 sec (+23% of the mean) -- maximum

Number of outlier (out of 1.30 sec..2.01 sec): 1
```